refactor `storage` package paths
- move version control root to subdirectory
- rename `file_system` to dirdict
- rename `FileSystemWrapperDict.py` to `DirDict.py`
