+++ dirdict/functions.py
