--- requirements.txt
@@ -1,3 +1,0 @@
-requests
-html5lib
-bs4
